add steps to recipe

+++ app/src/main/java/team3/recipefinder/model/Recipe.kt
@@ -6,27 +6,27 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "recipe")
 data class Recipe (
-    @PrimaryKey(autoGenerate = true) val id: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "name") val name: String
 )
 
 @Entity(tableName = "ingredient")
 data class Ingredient (
-    @PrimaryKey(autoGenerate = true) val id: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "name") val name: String
 )
 
 @Entity(tableName = "step")
 data class RecipeStep (
-    @PrimaryKey(autoGenerate = true) val id: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "description") val description: String
 )
 
 @Entity(tableName = "rel_recipe_ingredient")
 data class RelRecipeIngredient(
-    @PrimaryKey(autoGenerate = true) val id: Int,
-    @ColumnInfo(name = "recipeId") val recipeId: Int,
-    @ColumnInfo(name = "ingredientId") val ingredientId: Int
+    @PrimaryKey(autoGenerate = true) val id: Long,
+    @ColumnInfo(name = "recipeId") val recipeId: Long,
+    @ColumnInfo(name = "ingredientId") val ingredientId: Long
 )
 
 @Entity(tableName = "rel_recipe_step")
